requirements.txt
Pillow==8.2.0
opencv-python==4.5.2.54